Fix Gradient issue; move now passing tests back
