+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/autodiff/samediff/SameDiff.java
@@ -1849,7 +1849,15 @@ public class SameDiff {
      * @return the gradient for this variable or null
      */
     public SDVariable getGradForVariable(String varName) {
-        return gradients.get(varName);
+        //TODO 2018/06/26 - Review this?
+        //Gradients are being placed in the inner "grad" function SameDiff instance, but not the outer one
+        // should they be synced and we just use the map in this instance?
+        if (gradients.containsKey(varName)) {
+            return gradients.get(varName);
+        } else if(sameDiffFunctionInstances.containsKey("grad") && sameDiffFunctionInstances.get("grad").gradients.containsKey(varName)){
+            return sameDiffFunctionInstances.get("grad").gradients.get(varName);
+        }
+        return null;
     }
