+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/autodiff/samediff/FailingSameDiffTests.java
@@ -195,33 +195,6 @@ public class FailingSameDiffTests {
     }
 
     @Test
-    public void testUpdatingAssociateFwd(){
-        OpValidationSuite.ignoreFailing();
-
-        SameDiff sd = SameDiff.create();
-        SDVariable in = sd.var("in", Nd4j.linspace(1,12,12).reshape(3,4));
-        SDVariable w = sd.var("w", Nd4j.linspace(1,20,20).reshape(4,5));
-        SDVariable out = sd.mmul(in, w);
-        SDVariable loss = out.std("out", true);
-
-        INDArray outArr = sd.execAndEndResult().dup();
-        sd.execBackwards();
-
-        Map<String,INDArray> origGrad = new HashMap<>();
-        origGrad.put("in", in.gradient().getArr().dup());
-        origGrad.put("w", w.gradient().getArr().dup());
-        origGrad.put("out", out.gradient().getArr().dup());
-
-        INDArray newIn = in.getArr().dup().muli(5);
-        in.setArray(newIn);
-
-        //check gradient function copy of array
-        SameDiff sdGrad = sd.getFunction("grad");
-        INDArray gradArrIn = sdGrad.getVariable("in").getArr();
-        assertEquals(newIn, gradArrIn);
-    }
-
-    @Test
     public void testDropout() {
         OpValidationSuite.ignoreFailing();
         SameDiff sd = SameDiff.create();
@@ -233,81 +206,6 @@ public class FailingSameDiffTests {
         SDVariable res = sd.dropout(input, p);
         assertArrayEquals(new long[]{2, 2}, res.getShape());
     }
-
-    @Test
-    public void testUpdatingGradientSimple() throws Exception {
-        OpValidationSuite.ignoreFailing();
-
-        SameDiff sd = SameDiff.create();
-        SDVariable in = sd.var("in", Nd4j.linspace(1,12,12).reshape(3,4));
-        SDVariable out = in.mul(2.0);
-        SDVariable loss = out.std("out", true);
-
-        INDArray outArr = sd.execAndEndResult().dup();
-        sd.execBackwards();
-
-        SameDiff sdGrad = sd.getFunction("grad");
-
-        Field f = SameDiff.class.getDeclaredField("variableNameToArr");
-        f.setAccessible(true);
-        Map<String,INDArray> before = (Map<String, INDArray>) f.get(sdGrad);
-        for(Map.Entry<String,INDArray> e : before.entrySet()){
-            System.out.println(e.getKey());
-            System.out.println(e.getValue());
-        }
-
-        Map<String,INDArray> origGrad = new HashMap<>();
-        origGrad.put("in", in.gradient().getArr().dup());
-        origGrad.put("out", out.gradient().getArr().dup());
-
-        double stdBefore = in.getArr().stdNumber().doubleValue();
-        in.getArr().assign(Nd4j.rand(in.getArr().shape()));
-        double stdAfter = in.getArr().stdNumber().doubleValue();
-        System.out.println("Before vs. after: " + stdBefore + ", " + stdAfter);
-        INDArray outArr2 = sd.execAndEndResult();
-        sd.execBackwards();
-
-        assertNotEquals(outArr, outArr2);
-
-        System.out.println("=======================================================");
-        Map<String,INDArray> after = (Map<String, INDArray>) f.get(sdGrad);
-        for(Map.Entry<String,INDArray> e : after.entrySet()){
-            System.out.println(e.getKey());
-            System.out.println(e.getValue());
-        }
-
-        //Ensure gradients are also changed:
-        assertNotEquals(origGrad.get("in"), in.gradient().getArr());
-        assertNotEquals(origGrad.get("out"), out.gradient().getArr());
-    }
-
-
-    @Test
-    public void testUpdatingInplaceFwd(){
-        OpValidationSuite.ignoreFailing();
-
-        SameDiff sd = SameDiff.create();
-        SDVariable in = sd.var("in", Nd4j.linspace(1,12,12).reshape(3,4));
-        SDVariable w = sd.var("w", Nd4j.linspace(1,20,20).reshape(4,5));
-        SDVariable out = sd.mmul(in, w);
-        SDVariable loss = out.std("out", true);
-
-        INDArray outArr = sd.execAndEndResult().dup();
-        sd.execBackwards();
-
-        Map<String,INDArray> origGrad = new HashMap<>();
-        origGrad.put("in", in.gradient().getArr().dup());
-        origGrad.put("w", w.gradient().getArr().dup());
-        origGrad.put("out", out.gradient().getArr().dup());
-
-        in.getArr().muli(5);
-
-        //check gradient function copy of array
-        SameDiff sdGrad = sd.getFunction("grad");
-        INDArray gradArrIn = sdGrad.getVariable("in").getArr();
-        assertEquals(in.getArr(), gradArrIn);
-    }
-
 
     @Test
     public void testExecutionDifferentShapesDynamicCustom(){
@@ -342,73 +240,4 @@ public class FailingSameDiffTests {
         assertArrayEquals(new long[]{3,5}, out2.shape());
     }
 
-
-    @Test
-    public void testUpdatingGradient(){
-        OpValidationSuite.ignoreFailing();
-        Nd4j.getRandom().setSeed(12345);
-
-        SameDiff sd = SameDiff.create();
-        SDVariable in = sd.var("in", Nd4j.linspace(1,12,12).reshape(3,4));
-        SDVariable w = sd.var("w", Nd4j.linspace(1,20,20).reshape(4,5));
-        SDVariable out = sd.mmul(in, w);
-        SDVariable loss = out.std("out", true);
-
-        INDArray outArr = sd.execAndEndResult().dup();
-        sd.execBackwards();
-
-        Map<String,INDArray> origGrad = new HashMap<>();
-        origGrad.put("in", in.gradient().getArr().dup());
-        origGrad.put("w", w.gradient().getArr().dup());
-        origGrad.put("out", out.gradient().getArr().dup());
-
-        in.getArr().assign(Nd4j.rand(in.getArr().shape()));
-        INDArray outArr2 = sd.execAndEndResult();
-        sd.execBackwards();
-
-        assertNotEquals(outArr, outArr2);
-
-        //Ensure gradients are also changed:
-        assertNotEquals(origGrad.get("in"), in.gradient().getArr());
-        assertNotEquals(origGrad.get("w"), w.gradient().getArr());
-        assertNotEquals(origGrad.get("out"), out.gradient().getArr());
-    }
-
-
-    @Test
-    public void testExternalErrorsSimple(){
-        OpValidationSuite.ignoreFailing();
-        INDArray externalGrad = Nd4j.linspace(1,12,12).reshape(3,4);
-
-        SameDiff sd = SameDiff.create();
-        SDVariable var = sd.var("var", externalGrad);
-        SDVariable out = var.mul("out", 0.5);
-
-        Map<String,INDArray> gradMap = new HashMap<>();
-        gradMap.put("out", externalGrad);
-        ExternalErrorsFunction fn = sd.f().externalErrors(out);
-        //new ExternalErrorsFunction(sd, Collections.singletonList(out), gradMap);
-
-        fn.updateVariable("out", externalGrad);
-        sd.execAndEndResult();
-        sd.execBackwards();
-
-        INDArray gradOut = out.getGradient().getArr();
-        INDArray gradVar = var.getGradient().getArr();
-
-        assertEquals(externalGrad, gradOut);
-        assertEquals(externalGrad.mul(0.5), gradVar);
-
-        //Now, update and execute again:
-        externalGrad = Nd4j.linspace(1,12,12).reshape(3,4).muli(10);
-        fn.updateVariable("out", externalGrad);
-
-        sd.execBackwards();
-
-        gradOut = out.getGradient().getArr();
-        gradVar = var.getGradient().getArr();
-
-        assertEquals(externalGrad, gradOut);
-        assertEquals(externalGrad.mul(0.5), gradVar);
-    }
 }
